Move the backoff test to its own test module
